diarize/requirements.txt
pyannote.audio>=3.1
fastapi>=0.104
uvicorn>=0.24
torch>=2.0
torchaudio>=2.0
